% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{.import_jupytext}
\alias{.import_jupytext}
\title{Import the jupytext Python library}
\usage{
.import_jupytext(proc)
}
\arguments{
\item{proc}{A process object generated by \code{\link[basilisk:basiliskStart]{basilisk::basiliskStart()}}}
}
\value{
\code{NULL}, invisibly. The jupytext module is stored within the \code{proc} process object and can
be passed to subsequent \code{basiliskRun()} calls by setting \code{persist = TRUE}, passing a \code{store}
argument to \code{fun} and accesing the module with \code{store$jupytext}.
}
\description{
Load the jupytext Python library in an existing basilisk process.
}
\keyword{internal}
