% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jupytextR-package.R
\docType{package}
\name{jupytextR-package}
\alias{jupytextR}
\alias{jupytextR-package}
\title{jupytextR: R wrapper for jupytext}
\description{
Convert between Jupyter notebooks, [R]Markdown documents and scripts (R, Python, Julia, ...), using the jupytext Python package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/milanmlft/jupytextR}
  \item \url{https://milanmlft.github.io/jupytextR/}
  \item Report bugs at \url{https://github.com/milanmlft/jupytextR/issues}
}

}
\author{
\strong{Maintainer}: Milan Malfait \email{milan.malfait94@gmail.com} (\href{https://orcid.org/0000-0001-9144-3701}{ORCID})

}
\keyword{internal}
