% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_ipynb}
\alias{read_ipynb}
\title{Read ipynb files using jupytext's read() method}
\usage{
read_ipynb(input)
}
\arguments{
\item{input}{File path to the input file to be read.}
}
\value{
An R list containing the notebook data.
}
\description{
Calls the \code{jupytext.read} method from the
\href{https://jupytext.readthedocs.io/en/latest/using-library.html}{\code{jupytext} API}.
The ipynb data is loaded in memory as an R list, converted from the underlying JSON data of the
notebook.
}
\keyword{internal}
