% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{.import_jupytext}
\alias{.import_jupytext}
\title{Import the jupytext Python library}
\usage{
.import_jupytext(proc)
}
\arguments{
\item{proc}{A process object generated by \code{\link[basilisk:basiliskStart]{basilisk::basiliskStart()}}}
}
\value{
An R wrapper for the jupytext module
}
\description{
Load the jupytext Python library in an existing basilisk process.
}
\keyword{internal}
